feat: 新增以下接口： 1. 版本更新检测（用户客户端） 2. 版本发布 （管理员） 3. 版本移除 （管理员） 4. 所有版本 （管理员）

+++ src/main/java/com/fangxiaoyun/springboot/myspringboot/table/Version.java
@@ -26,7 +26,7 @@ public class Version {
 
     // 最低不强制更新版本号
     @Column(name = "mini_version_code")
-    private int miniVersionCode;
+    private long miniVersionCode;
 
     // 更新提示标题
     @Column(name = "title")
